+++ packages/lumx-react/src/stories/controls/selectArgType.ts
@@ -1,10 +1,8 @@
-import isPlainObject from 'lodash/isPlainObject';
-
 export const getSelectArgType = <E>(
     options: Array<E> | Record<string, E>,
     type: 'select' | 'inline-radio' = 'inline-radio',
 ) => ({
     control: { type },
     options: Object.values(options),
-    mapping: isPlainObject(options) ? options : undefined,
+    mapping: !Array.isArray(options) ? options : undefined,
 });