+++ CHANGELOG.md
@@ -11,9 +11,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 -   Storybook dark theme switcher: improve label of dark/light modes
 -   Storybook `withCombinations()`: better section heading, exclude combinations
--   Reworked Button storybook stories
+-   Storybook `withTheming()`: new decorator to demonstrate CSS variables on stories
+-   Storybook Button:
     -    Cleaned up stories argTypes and controls
     -    Rework stories content with more combinations
+    -    Add theming story
 
 ## [3.9.5][] - 2024-11-06